requirements.txt
Flask==2.0.1
numpy==1.19.2
pandas==1.2.5
PyYAML==5.4.1
scikit_learn==0.24.2
tensorflow==2.5.0
tqdm==4.61.1
